fix: update package dependencies to use catalog references
