+++ package.json
@@ -30,7 +30,7 @@
     "turbo": "catalog:",
     "typescript": "catalog:",
     "vitest": "catalog:",
-    "msw": "^2.7.3"
+    "msw": "catalog:"
   },
   "pnpm": {
     "onlyBuiltDependencies": [
