+++ packages/cli/package.json
@@ -38,16 +38,14 @@
     "farver": "catalog:",
     "fs-extra": "catalog:",
     "semver": "catalog:",
-    "yargs": "catalog:",
-    "yargs-parser": "^21.1.1",
+    "yargs-parser": "catalog:",
     "zod": "catalog:"
   },
   "devDependencies": {
     "@luxass/eslint-config": "catalog:",
     "@types/fs-extra": "catalog:",
     "@types/semver": "catalog:",
-    "@types/yargs": "catalog:",
-    "@types/yargs-parser": "^21.0.3",
+    "@types/yargs-parser": "catalog:",
     "eslint": "catalog:",
     "tsup": "catalog:",
     "typescript": "catalog:",
