+++ pnpm-workspace.yaml
@@ -8,12 +8,15 @@ catalog:
   "@types/node": ^22.10.0
   "@types/semver": ^7.5.8
   "@types/yargs": ^17.0.33
+  "@types/yargs-parser": ^21.0.3
   "@vitest/coverage-v8": ^3.0.7
   "@vitest/ui": ^3.0.7
+  apache-autoindex-parse: ^0.2.0
   defu: ^6.1.4
   eslint: ^9.21.0
   farver: ^0.4.1
   fs-extra: ^11.3.0
+  msw: ^2.7.3
   semver: ^7.7.1
   tsup: ^8.4.0
   turbo: ^2.4.4
@@ -21,4 +24,5 @@ catalog:
   vitest: ^3.0.7
   vitest-testdirs: ^2.1.1
   yargs: ^17.7.2
+  yargs-parser: ^21.1.1
   zod: ^3.24.2
